refactor(runtime): separate workflow scheduling and postconditions
