+++ src/crewplane/runtime/execution/workflow/__init__.py
@@ -1,5 +1,6 @@
 """Workflow execution facade."""
 
-from .orchestration import execute_workflow, wait_for_completed_nodes
+from .orchestration import execute_workflow
+from .scheduling import wait_for_completed_nodes
 
 __all__ = ["execute_workflow", "wait_for_completed_nodes"]
